requirements.txt
flask
paho-mqtt
python-dotenv
uwsgi
schedule